+++ README.md
@@ -10,56 +10,12 @@ A Flutter package for seamless integration with Azure authentication and Silent 
 - **Useful for Offline Applications:** Silent login functionality is particularly useful for offline applications where continuous user authentication may not be possible.
 - **Token Management:** Automatically manages token expiration and refreshing.
 
-### Explanation:
-
-1. **First-Time Login:**
-   - When the user opens the app for the first time, call `AzureAuth->Login()` to open the authentication page.
-   - Upon successful login, the token is securely stored.
-   
-2. **Token Retrieval:**
-   - When the app needs the token, call `AzureAuth->getAccessToken()`.
-   - The token's validity is checked; if expired, it is automatically refreshed.
-   
-3. **Silent Login:**
-   - Use `AzureAuth->silentLogin()` when the app is opened again.
-   - The stored token is validated in the background without redirecting the user to the authentication page.
-   - If the token is invalid or expired, an error is thrown, allowing developers to redirect the user to the login screen.
-   
-4. **Logout:**
-   - Call `AzureAuth->logout()` to clear all locally stored token and user data.
-   - Sends a request to the OpenID Provider to log out the End-User.
-
-### Documentation
-
-Explore the documentation to learn more.
-
-
-## Getting started
-
-To use this package, add the following dependency to your `pubspec.yaml` file:
-
-```yaml
-dependencies:
-  azure_silent_auth: ^1.0.0
-```
-
-Then, run:
-
-```bash
-flutter pub get
-```
-### Prerequisites
-
-Flutter: Your app should use Flutter framework.
-Dart: Your app should use Dart programming language.
-
-### Platform
-
-Tested in Windows and iOS, Should work on Android too.
-
 ## Usage
 
 ```dart
+import 'package:azure_silent_auth/authenticator/default_authenticator.dart';
+import 'package:azure_silent_auth/azure_silent_auth.dart';
+
   AzureAuth _azureAuth = AzureAuth(
     authenticatorProvider: DefaultAuthenticator(
       'https://login.microsoftonline.com/{tenant-id}/',
@@ -70,7 +26,29 @@ Tested in Windows and iOS, Should work on Android too.
     ),
   );
 ```
-Please refer main.dart in the example app to see how this instance can be used in different scenarios.
+
+1. **First-Time Login:**
+   - When the user opens the app for the first time, call `await _azureAuth.login();` to open the authentication page.
+   - Upon successful login, the token is securely stored.
+   
+2. **Token Retrieval:**
+   - When the app needs the token, call `await _azureAuth.getAccessToken();`.
+   - The token's validity is checked; if expired, it is automatically refreshed.
+   
+3. **User Info Retrieval:**
+   - When the app needs the loggedin users details, call `_azureAuth.getUserInfo();`.
+
+4. **Silent Login:**
+   - Use `await _azureAuth.silentLogin();` when the app is opened again.
+   - The stored token is validated in the background without redirecting the user to the authentication page.
+   - If the token is invalid or expired, an error is thrown, allowing developers to redirect the user to the login screen.
+   
+5. **Logout:**
+   - Call `_azureAuth.logout();` to clear all locally stored token and user data.
+   - Sends a request to the OpenID Provider to log out the End-User.
+   
+Please refer example app to see how this instance can be used in different scenarios.
+
 ## Additional information
 
 - Contributions: Contributions are welcome! Open an issue or submit a pull request.
